requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
rich>=10.0.0
pyperclip>=1.8.2